monitor/holdings.py
from futu import *
import logging

def get_holdings(trade_ip=None, trade_port=None, use_paper=False):
    """
    可替换为真实持仓：
      trade_ctx = OpenFutureTradeContext(...) 或 OpenUSTradeContext(...)
      ret, data = trade_ctx.position_list_query()
    目前先返回空集合，表示无持仓。
    """
    logging.info("使用占位持仓（未接入交易上下文）")
    return set()
