requirements.txt
futu-api
numpy
pandas
TA-Lib
PyWavelets
scikit-learn
scipy
schedule
requests
python-dateutil
matplotlib